README: describe setting known_hosts, add remote server steps
Just having gone through the process of installing, configuring
and debugging sidedoor, this commits tries to make a few spots
in the README clearer to make it easier for the next person.

It also mentions that you should set "ClientAliveInterval 30"
on the server side. This is a footgun that I hit when testing
sidedoor: Without it, a hung ssh session can block the port
and the remote server for a long time (days? hours?), making
the tunnel unusable.

Fixes https://github.com/daradib/sidedoor/issues/3